ganesha_mgr gets a new option to show mounted POSIX filesystems
The "show posix_fs" option is added to the ganesha_mgr utility.
It will show the major and minor device id's of the mounted
filesystems.

The "show" subcommand can also be used to display exports and
clients via "show exports" and "show clients" respectively.

The older interfaces of show_exports and show_client are still
retained.